Test implementation of an excel parser
- created an implementation of the parsing of feit.xls file in the
ReadExcelTest class. It prints a content of a certain sheet to a text
file;
- Apache POI lib is used to read the content of the excel file. It seems
that there are some bugs in this lib during reading a file: some cells
are read incorrectly.

Note: to copy the feit.xls file to an emulator open the 'DDMS'
perspective and push this file to the data/data/ua.chntu.sheduler/files
directory.